Load project specific ssh keys first.
Problem: We want to set up buildkite CI in the private project of the
another organization and we cannot invite serokell-bot to it without
additional expenses.

Solution: Change the order in which ssh keys are fetched. Load ssh key
from the `serokell-private/<project-name>/private_ssh_key` first, so
that it can be checked before the global key.